feat: use slugs for public routes
