+++ client/components/News.vue
@@ -5,8 +5,8 @@
         <div class="news__left-items">
           <!-- class="news__left-item" -->
           <div v-for="item in news.news" :key="item.newsId">
-            <!-- @click="$router.push(localeLocation(`/news/${item.newsId}`))" -->
-            <a :href="`/${$i18n.locale}/news/${item.newsId}`" class="news__left-item">
+            <!-- @click="$router.push(localeLocation(`/news/${item.slug}`))" -->
+            <a :href="`/${$i18n.locale}/news/${item.slug}`" class="news__left-item">
               <div class="news__left-item-image">
                 <img :src="`${imageURL}${item?.image}`" alt="" />
               </div>
@@ -26,7 +26,7 @@
           {{ $t('seeAll') }}
         </button>
       </div>
-      <a :href="`${$i18n.locale}/news/${news?.mainNews?.newsId}`">
+      <a :href="`${$i18n.locale}/news/${news?.mainNews?.slug}`">
         <div class="news__center">
           <div class="news__center-image">
             <img :src="`${imageURL}${news?.mainNews?.image}`" alt="" />
