+++ client/pages/products-services/_slug.vue
@@ -1,114 +1,79 @@
 <template>
-  <div class="projects-id" ref="aos">
-    <div class="projects-id__container">
-      <div class="projects-id__arrow">
+  <div class="products" ref="aos">
+    <div class="products__container">
+      <div class="products__arrow">
         <base-icon
           icon="arrowLeft"
           class="project-icon"
-          @clicked="$router.push(localeLocation(`/${$i18n.locale}/projects`))"
+          @clicked="$router.back(localeLocation(-1))"
         />
       </div>
-      <div class="projects-id__project">
-        <div class="projects-id__project-image" style="margin-right: 15px">
-          <!-- <img src="@/assets/img/mennan.svg" alt="" /> -->
-          <img
-            :src="`${imageURL}${project.cover}`"
-            alt=""
-            @click="openModal(`${imageURL}${project.cover}`)"
-          />
+      <div class="products__project">
+        <div class="products__project-image">
+          <img :src="`${imageURL}${product?.logo}`" alt="" />
         </div>
-        <div class="projects-id__project-content">
-          <p class="projects-id__project-text">
-            <span class="projects-id__project-span">{{ $t('company') }}:</span>
-            <span>
-              {{ project?.[translator('company')] }}
-            </span>
+        <div class="products__project-content">
+          <p class="products__project-text">
+            <!-- <span>Name:</span> -->
+            <!-- <span>{{ translateName(product) }}</span> -->
           </p>
-          <p class="projects-id__project-text">
-            <span class="projects-id__project-span">{{ $t('engineeringPeriod') }}:</span>
-            <span>
-              {{
-                new Date(project?.workDate).toLocaleString(translateLanguage(project), {
-                  month: 'long',
-                })
-              }}, {{ new Date(project?.workDate).getFullYear() }}
-            </span>
-            <span
-              v-if="
-                project?.endDate &&
-                new Date(project?.workDate).getMonth() !== new Date(project?.endDate).getMonth()
-              "
-            >
-              -
-              {{
-                new Date(project?.endDate).toLocaleString(translateLanguage(project), {
-                  month: 'long',
-                })
-              }}, {{ new Date(project?.endDate).getFullYear() }}
-            </span>
-          </p>
-          <p class="projects-id__project-text">
-            <span v-html="project[translator(`name`)]" style="font-weight: 600"></span>
-          </p>
+          <!-- <p class="products__project-text">
+            <span>Type:</span>
+            <span>{{ product?.type }}</span>
+          </p> -->
         </div>
       </div>
-      <div class="projects-id__description-wrapper">
-        <h1 class="projects-id__description-title">{{ $t('description') }}</h1>
+      <p style="color: #183a60; font-size: large; font-weight: 600; margin-bottom: 10px">
+        {{ product[translator('name')] }}
+      </p>
+      <div class="products__description-wrapper">
+        <h1 class="products__description-title">{{ $t('description') }}</h1>
+
         <p
-          class="projects-id__description-description"
-          v-html="project?.[translator(`description`)]"
+          class="products__description-description inner_description"
+          style="line-height: 2em; text-align: justify; text-indent: 2em"
+          v-html="product?.[translator(`content`)]"
         ></p>
+        <!-- v-html="translateContent(item)" -->
       </div>
-      <div class="projects-id__images-wrapper">
-        <!-- <h1 class="projects-id__images-title">{{ $t("photo") }}</h1> -->
-        <div class="projects-id__images-row" ref="images">
+      <!-- <div class="products__images-wrapper">
+        <h1 class="products__images-title">{{ $t("imagesText") }}</h1>
+        <div class="products__images-row" ref="images">
           <div
-            class="projects-id__images-img"
-            v-for="photo in project?.images"
+            class="products__images-img"
+            v-for="photo in product?.images"
             :key="photo"
-            @click="openModal(`${imageURL}${photo}`)"
           >
             <img :src="`${imageURL}${photo}`" alt="" />
           </div>
         </div>
-      </div>
-      <ImagePreviewModal
-        :imageUrl="selectedImage"
-        :isVisible="isModalVisible"
-        @close="closeModal"
-      />
+      </div> -->
     </div>
   </div>
 </template>
 
 <script>
-import { GET_PROJECT_ONE } from '@/api/home.api';
+import { GET_PRODUCTS_ONE } from '@/api/home.api';
 import translate from '@/mixins/translate';
 import { mapGetters } from 'vuex';
-import ImagePreviewModal from '~/components/ImagePreviewModal.vue';
 
 export default {
-  components: {
-    ImagePreviewModal,
-  },
   mixins: [translate],
   computed: {
     ...mapGetters(['imageURL']),
   },
   data() {
     return {
-      project: {
+      product: {
         images: {
           type: Array,
           default: () => [],
         },
       },
-      isModalVisible: false,
-      selectedImage: '',
     };
   },
   async mounted() {
-    await this.fetchProject();
+    await this.fetchProducts();
     if (this.$refs.aos) {
       const options =
         {
@@ -131,34 +96,24 @@ export default {
   },
 
   methods: {
-    async fetchProject() {
+    async fetchProducts() {
       try {
-        const { data, statusCode } = await GET_PROJECT_ONE({
-          data: {
-            projectId: this.$route.params.id,
-          },
+        const { data, statusCode } = await GET_PRODUCTS_ONE({
+          slug: this.$route.params.slug,
         });
         if (statusCode) {
-          this.project = data || {};
+          this.product = data || {};
         }
       } catch (error) {
         console.error(error);
       }
-    },
-    openModal(imageUrl) {
-      this.selectedImage = imageUrl;
-      this.isModalVisible = true;
-    },
-    closeModal() {
-      this.isModalVisible = false;
-      this.selectedImage = '';
     },
   },
 };
 </script>
 
 <style lang="scss" scoped>
-.projects-id {
+.products {
   padding: 120px 0;
   @media (max-width: 767px) {
     padding: 30px 0;
@@ -198,22 +153,18 @@ export default {
     display: flex;
     justify-content: space-between;
     align-items: center;
+    max-width: 980px;
     margin-bottom: 30px;
   }
 
   &__project-image {
-    flex: 0 0 25%;
-    height: 155px;
+    max-width: 180px;
+    height: 100px;
     img {
       width: 100%;
       height: 100%;
-      object-fit: cover;
+      object-fit: contain;
       object-position: center;
-      border-radius: 4px;
-    }
-    @media (max-width: 479px) {
-      flex: 0 0 50%;
-      height: 130px;
     }
   }
 
@@ -221,7 +172,6 @@ export default {
     display: flex;
     flex-direction: column;
     gap: 8px;
-    width: 750px;
   }
 
   &__project-text {
@@ -229,13 +179,11 @@ export default {
     font-size: 16px;
     font-weight: 400;
     line-height: normal;
-    text-transform: initial;
+    text-transform: capitalize;
   }
 
   &__project-span {
     margin-right: 4px;
-    font-weight: 600;
-    text-transform: initial;
   }
 
   &__description-wrapper {
@@ -247,7 +195,7 @@ export default {
     font-size: 30px;
     font-weight: 400;
     letter-spacing: 0.45px;
-    // text-transform: capitalize;
+    text-transform: capitalize;
     padding-bottom: 7px;
     position: relative;
     display: inline-block;
@@ -276,12 +224,16 @@ export default {
     }
   }
 
+  .inner_description > ul {
+    list-style-type: unset !important;
+  }
+
   &__description-description {
+    list-style-type: unset;
     color: #000;
     font-size: 17px;
-    text-indent: 2em;
     font-weight: 500;
-    line-height: 2em;
+    line-height: normal;
     @media (max-width: 767px) {
       font-size: 15px;
     }
@@ -332,16 +284,13 @@ export default {
     overflow-x: auto;
     max-width: 100%;
     padding-bottom: 10px;
-
-    @media (min-width: 767px) {
+    transition: 1s all;
+    transform: translateY(120px);
+    opacity: 0;
+    &.aos {
+      opacity: 1;
+      transform: translateY(0px);
       transition: 1s all;
-      transform: translateY(120px);
-      opacity: 0;
-      &.aos {
-        opacity: 1;
-        transform: translateY(0px);
-        transition: 1s all;
-      }
     }
     @media (max-width: 479px) {
       gap: 16px;
@@ -352,8 +301,10 @@ export default {
     flex: 0 0 25%;
     height: 155px;
     img {
-      width: 100%;
-      height: 100%;
+      max-width: 100%;
+      width: auto;
+      max-height: 100%;
+      height: auto;
       object-fit: cover;
       object-position: center;
       border-radius: 4px;
