+++ client/pages/projects/index.vue
@@ -11,8 +11,8 @@
       </div>
       <div class="projects__items" ref="images">
         <div v-for="project in projects.rows" :key="project.projectId" class="projects-item">
-          <!-- @click="openInNewTab(project.projectId)" -->
-          <a :href="`projects/${project.projectId}`" class="projects-item">
+          <!-- @click="openInNewTab(project.slug)" -->
+          <a :href="`${project.slug}`" class="projects-item">
             <div class="projects-item__image">
               <img :src="`${imageURL}${project?.cover}`" alt="" />
             </div>
@@ -81,8 +81,8 @@ export default {
   },
 
   methods: {
-    openInNewTab(projectId) {
-      this.$router.push(`/projects/${projectId}`);
+    openInNewTab(slug) {
+      this.$router.push(`/projects/${slug}`);
     },
 
     async fetchProjects() {
