+++ server/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { FileHelper } from './file-delete.util';
 export { RequestWithUser } from './request-with-user';
 export { responseInterceptor } from './response.interceptor';
+export { slugify, slugWithId } from './slug.util';
