+++ server/src/utils/slug.util.ts
@@ -1,0 +1,14 @@
+export function slugify(text: string): string {
+  return text
+    .toString()
+    .normalize('NFKD')
+    .replace(/[\u0300-\u036F]/g, '')
+    .toLowerCase()
+    .trim()
+    .replace(/[^a-z0-9]+/g, '-')
+    .replace(/^-+|-+$/g, '');
+}
+
+export function slugWithId(text: string, id: string): string {
+  return `${slugify(text)}_${id}`;
+}
